Move post types to config folder, update refs
