+++ lib/wordpress/_config/postTypes.js
