+++ lib/wordpress/_global/getPostTypeArchive.js
@@ -4,7 +4,7 @@ import {initializeWpApollo} from '@/lib/wordpress/connector'
 import getMenus from '@/lib/wordpress/menus/getMenus'
 import queryPostsArchive from '@/lib/wordpress/posts/queryPostsArchive'
 import queryTeamsArchive from '@/lib/wordpress/teams/queryTeamsArchive'
-import {postTypes} from './postTypes'
+import {postTypes} from '@/lib/wordpress/_config/postTypes'
 
 // Define SEO for archives.
 export const archiveQuerySeo = {
