+++ lib/wordpress/_global/getPostTypeById.js
@@ -1,7 +1,7 @@
 import queryPageById from '@/lib/wordpress/pages/queryPageById'
 import queryPostById from '@/lib/wordpress/posts/queryPostById'
 import queryTeamById from '@/lib/wordpress/teams/queryTeamById'
-import {isHierarchicalPostType} from './postTypes'
+import {isHierarchicalPostType} from '@/lib/wordpress/_config/postTypes'
 import processPostTypeQuery from './processPostTypeQuery'
 
 /**
