+++ lib/wordpress/_global/getPostTypeStaticPaths.js
@@ -1,6 +1,10 @@
 import {initializeWpApollo} from '@/lib/wordpress/connector'
+import {
+  isHierarchicalPostType,
+  isValidPostType,
+  postTypes
+} from '@/lib/wordpress/_config/postTypes'
 import {gql} from '@apollo/client'
-import {isHierarchicalPostType, isValidPostType, postTypes} from './postTypes'
 
 /**
  * Retrieve static paths by post type.
