+++ lib/wordpress/_global/getPostTypeTaxonomyArchive.js
@@ -3,7 +3,7 @@ import queryPostsByCategory from '@/lib/wordpress/categories/queryPostsByCategor
 import {initializeWpApollo} from '@/lib/wordpress/connector'
 import getMenus from '@/lib/wordpress/menus/getMenus'
 import queryPostsByTag from '@/lib/wordpress/tags/queryPostsByTag'
-import {postTypes} from './postTypes'
+import {postTypes} from '@/lib/wordpress/_config/postTypes'
 
 /**
  * Retrieve post taxnomy archive.
